+++ src/main/java/com/lizzy/demo/login/mapper/LoginUserMappers.java
@@ -9,8 +9,8 @@ import org.apache.ibatis.annotations.Select;
 @Mapper//注入容器
 public interface LoginUserMappers {
     @Insert("insert into user(Username,Password,UserType,UserID) values(#{Username},#{Password},#{UserType},#{UserID})")
-    void insert(@Param("Username")String Username,@Param("UserType")int UserType
-            ,@Param("Password")String Password,@Param("UserID")long UserID);
+    void insert(@Param("Username")String Username,@Param("UserType")Integer UserType
+            ,@Param("Password")String Password,@Param("UserID")Integer UserID);
     @Select("select * from user where Username = #{Username}  and UserType = #{UserType}")
     User selectByUsername(@Param("Username")String Username,@Param("UserType")int UserType);
     @Select("select * from user where UserID = #{UserID}")
